Pass environmentObject and remove coordinator
