+++ BookPlayer/Library/ItemDetails Screen/Views/ItemDetailsArtworkSectionView.swift
@@ -15,7 +15,7 @@ struct ItemDetailsArtworkSectionView: View {
   /// Callback for action handler
   var actionHandler: () -> Void
   /// Theme view model to update colors
-  @StateObject var themeViewModel = ThemeViewModel()
+  @EnvironmentObject var themeViewModel: ThemeViewModel
 
   var body: some View {
     Section(
