+++ BookPlayer/Profile/Profile Screen/ProfileView.swift
@@ -17,18 +17,14 @@ struct ProfileView<Model: ProfileViewModelProtocol>: View {
     GeometryReader { geometryProxy in
       ScrollView {
         VStack(spacing: Spacing.M) {
-          ProfileCardView(
-            account: $viewModel.account,
-            themeViewModel: themeViewModel
-          )
-          .onTapGesture {
-            viewModel.showAccount()
-          }
-          .padding([.top, .trailing, .leading], Spacing.S)
+          ProfileCardView(account: $viewModel.account)
+            .onTapGesture {
+              viewModel.showAccount()
+            }
+            .padding([.top, .trailing, .leading], Spacing.S)
 
           ProfileListenedTimeView(
-            formattedListeningTime: $viewModel.totalListeningTimeFormatted,
-            themeViewModel: themeViewModel
+            formattedListeningTime: $viewModel.totalListeningTimeFormatted
           )
 
           Spacer()
@@ -50,6 +46,7 @@ struct ProfileView<Model: ProfileViewModelProtocol>: View {
         .frame(maxWidth: .infinity, minHeight: geometryProxy.size.height)
       }
     }
+    .environmentObject(themeViewModel)
   }
 }
